requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pydantic==2.5.0
python-dotenv==1.0.0
cors==1.0.1
vercel-blob==0.4.2
aiohttp==3.9.1
groq==0.4.1
